requirements.txt
torch
numpy
tiktoken
datasets
tqdm
pyyaml
scikit-learn
wandb>=0.16
